updated pagination functionality to be more closely tied to the pagination component and setup values passed as contextual for it
